restructuring code to keep autocompletion functions out of extension.js
